Moved the loop stuff into a seperate class
